+++ tests/testthat/test-add_cols.R
@@ -125,7 +125,7 @@ test_that(".add_hospitalisation works as expected", {
   linelist <- .add_hospitalisation(
     .data = ll,
     onset_to_hosp = onset_to_hosp,
-    hosp_rate = 0.5
+    hosp_risk = 0.5
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$hospitalisation, type = "double")
@@ -141,7 +141,7 @@ test_that(".add_hospitalisation works as expected with different parameter", {
   linelist <- .add_hospitalisation(
     .data = ll,
     onset_to_hosp = onset_to_hosp,
-    hosp_rate = 0.1
+    hosp_risk = 0.1
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$hospitalisation, type = "double")
@@ -152,17 +152,17 @@ test_that(".add_hospitalisation works as expected with different parameter", {
   )
 })
 
-test_that(".add_hospitalisation works as expected with age-strat rates", {
+test_that(".add_hospitalisation works as expected with age-strat risks", {
   ll <- readRDS(file.path("testdata", "pre_hospitalisation.rds"))
-  age_dep_hosp_rate <- data.frame(
+  age_dep_hosp_risk <- data.frame(
     min_age = c(1, 5, 80),
     max_age = c(4, 79, 90),
-    rate = c(0.1, 0.05, 0.2)
+    risk = c(0.1, 0.05, 0.2)
   )
   linelist <- .add_hospitalisation(
     .data = ll,
     onset_to_hosp = onset_to_hosp,
-    hosp_rate = age_dep_hosp_rate
+    hosp_risk = age_dep_hosp_risk
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$hospitalisation, type = "double")
@@ -178,8 +178,8 @@ test_that(".add_deaths works as expected", {
   linelist <- .add_deaths(
     .data = ll,
     onset_to_death = onset_to_death,
-    hosp_death_rate = 0.5,
-    non_hosp_death_rate = 0.5
+    hosp_death_risk = 0.5,
+    non_hosp_death_risk = 0.5
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$deaths, type = "double")
@@ -195,8 +195,8 @@ test_that(".add_deaths works as expected with different parameter", {
   linelist <- .add_deaths(
     .data = ll,
     onset_to_death = onset_to_death,
-    hosp_death_rate = 0.9,
-    non_hosp_death_rate = 0.1
+    hosp_death_risk = 0.9,
+    non_hosp_death_risk = 0.1
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$deaths, type = "double")
@@ -207,23 +207,23 @@ test_that(".add_deaths works as expected with different parameter", {
   )
 })
 
-test_that(".add_deaths works as expected with age-strat rates", {
+test_that(".add_deaths works as expected with age-strat risks", {
   ll <- readRDS(file.path("testdata", "pre_death.rds"))
-  age_dep_hosp_death_rate <- data.frame(
+  age_dep_hosp_death_risk <- data.frame(
     min_age = c(1, 5, 80),
     max_age = c(4, 79, 90),
-    rate = c(0.1, 0.05, 0.2)
+    risk = c(0.1, 0.05, 0.2)
   )
-  age_dep_non_hosp_death_rate <- data.frame(
+  age_dep_non_hosp_death_risk <- data.frame(
     min_age = c(1, 5, 80),
     max_age = c(4, 79, 90),
-    rate = c(0.05, 0.025, 0.1)
+    risk = c(0.05, 0.025, 0.1)
   )
   linelist <- .add_deaths(
     .data = ll,
     onset_to_death = onset_to_death,
-    hosp_death_rate = age_dep_hosp_death_rate,
-    non_hosp_death_rate = age_dep_non_hosp_death_rate
+    hosp_death_risk = age_dep_hosp_death_risk,
+    non_hosp_death_risk = age_dep_non_hosp_death_risk
   )
   expect_s3_class(linelist, class = "data.frame")
   expect_type(linelist$deaths, type = "double")